feat: integrate ThemeSwitcherDialog into main component and remove WelcomePopup
